Remove CUDA `CMAKE_CXX_FLAGS*` passthrough again
This was shown not to be necessary if `CUDA_CXX_FLAGS*` is
initialized correctly.

It may happen that, when the first CMake configuration fails with
"CUDA not found" and you make CUDA available and try again without
cleaning the build directory, you end up with empty values for all
`CMAKE_CUDA_FLAGS* variables.