doc: Fix incorrect `Sphinx_EXECUTABLE` variable name
According to `FindSphinx.cmake` of this repository, the variable for the
Sphinx executable is called `Sphinx_EXECUTABLE` and not
`SPHINX_EXECUTABLE`.  As CMake variables are case sensitive [1], the
command `make sphinx` did not work and failed.

Furthermore, this commit fixes a Sphinx warning.  A space was missing
after an inline code block which meant that `ARCH_INDEPENDENT` was not
highlighted correctly.

[1] https://cmake.org/cmake/help/latest/manual/cmake-language.7.html#variables